% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCorrImage.R
\name{comp_images_corr}
\alias{comp_images_corr}
\title{Compute voxel-wise correlations between two groups of images}
\usage{
comp_images_corr(img_1, img_2, mask = NULL)
}
\arguments{
\item{img_1}{a matrix of multiple images}

\item{img_2}{a matrix of multiple images}

\item{mask}{a vector of mask taking logical values. The default is NULL.}

\item{coords}{a matrix of spatial coords of the images. The default is NULL.}
}
\value{
a vector of the correlation values
}
\description{
This function is to compute voxel-wise spatially varying correlation between two groups of 3D images
}
\examples{
set.seed(1000)
dim_img = c(10,10,10)
n = 50
grids <- lapply(1:3,function(i) seq(-round(dim_img[i]/2)+1,round(dim_img[i]/2),length=dim_img[i]))
coords <- expand.grid(grids)
num_voxels <- prod(dim_img)
img_1 <- array(rnorm(num_voxels*n),dim=c(num_voxels,n))
img_2 <- array(rnorm(num_voxels*n),dim=c(num_voxels,n))
cor_region <- create_sphere_mask(grids,radius=2)
cor_region_list <- array(c(cor_region),dim=dim(img_2))
img_2 <- ifelse(cor_region_list,img_1+rnorm(num_voxels*n,sd=0.5),img_2)
mask <- create_sphere_mask(grids,radius=4)
cor_map <- comp_images_corr(img_1,img_2,c(mask))
plot_3D_image_slices(array(cor_map,dim=dim_img),grids,c(-2,2))
}
\author{
Jian Kang <jiankang@umich.edu>
}
