% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCorrImage.R
\name{create_sphere_mask}
\alias{create_sphere_mask}
\title{Create sphere mask for image}
\usage{
create_sphere_mask(grids, center = rep(0, length = length(grids)), radius = 3)
}
\arguments{
\item{grids}{a list of grids for each dimenion of the image.}

\item{center}{the center of sphere. Default value is a vector of zeros.}

\item{radius}{a postive number indicates the size of the neighborhood.}
}
\value{
an array with logical values
}
\description{
Create sphere mask for image
}
\examples{
sphere_mask = create_sphere_mask(list(1:10,1:10,1:10))
}
\author{
Jian Kang <jiankang@umich.edu>
}
