% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCorrImage.R
\name{find_image_neighbors}
\alias{find_image_neighbors}
\title{Find the neighbors for each voxels in images}
\usage{
find_image_neighbors(grids, d = NULL, radius = 1)
}
\arguments{
\item{grids}{a vector of grids or a list of grids for each dimenion of the image.}

\item{d}{the dimension of the image. When \code{grids} is a vector, \code{d} should be specified. Default is NULL.}

\item{radius}{the size of neighborhood.}
}
\value{
a list object contains two matrices: \code{idx} and \code{dist}.
\code{idx} contains the neighbor indices of each voxel in each row.
\code{dist} contains the distances between neighbors and the voxel in each row.
}
\description{
Find the neighbors for each voxels in images
}
\examples{
nb <- find_image_neighbors(1:10,d=3)
}
\author{
Jian Kang <jiankang@umich.edu>
}
