% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCorrImage.R
\name{minimum_contrast_images}
\alias{minimum_contrast_images}
\title{Minimum contrast method estimate image correations}
\usage{
minimum_contrast_images(X, coords)
}
\arguments{
\item{X}{a matrix format of multiple images, where rows are different images in vector format.}

\item{coords}{voxel locations of images: each raw has the (x,y) coordinates for 2D images or (x, y, z) coordinates for 3D images.}
}
\value{
the rho value
}
\description{
This function estimate the rho parameter in the exponetial squared covariance kernel exp(-rho*||x - x'||^2)
}
\examples{
n = 10
coords = expand.grid(seq(-1,1,length=n),seq(-1,1,length=n))
rho0 = 0.5
dist_coords = dist(coords)
Sigma = exp(-rho0*as.matrix(dist_coords^1.99))
R <- chol(Sigma)
m = nrow(coords)
N = 20
X <- t(crossprod(R,matrix(rnorm(m*N),nrow=m,ncol=N)))
image(matrix(X[1,],n,n))
minimum_contrast_images(X,coords)

}
\author{
Jian Kang <jiankang@umich.edu>
}
