% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCorrImage.R
\name{neighbors_in_regions}
\alias{neighbors_in_regions}
\title{Create the neighbor indices on a group of subregions.}
\usage{
neighbors_in_regions(region_idx, nb_idx, nb_dist)
}
\arguments{
\item{region_idx}{a list of voxels indices for multiple regions. The length is the number of regions.}

\item{nb_idx}{matrix with the neighbor indices of each voxel in each row}

\item{nb_dist}{matrix with the distances between neighbors and the voxel in each row}
}
\value{
a list object contains of multiple list objects for neighbor indices and distances for different regions. For each region,
the list object contions two matrices: \code{idx} and \code{dist}.
\code{idx} contains the neighbor indices of each voxel in each row.
\code{dist} contains the distances between neighbors and the voxel in each row.
}
\description{
Create the neighbor indices on a group of subregions.
}
\examples{
dim_img = c(10,10,10)
grids <- lapply(1:3,function(i) seq(-round(dim_img[i]/2)+1,round(dim_img[i]/2),length=dim_img[i]))
nb <- find_image_neighbors(grids)
region_mask_3 <- create_sphere_mask(grids,radius=3)
region_mask_2 <- create_sphere_mask(grids,radius=2)
region_idx <- list(which(region_mask_3&(!region_mask_2)),which(region_mask_2))
nb_regions <- neighbors_in_regions(region_idx,nb$idx,nb$dist)
}
\author{
Jian Kang <jiankang@umich.edu>
}
