% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCorrImage.R
\name{plot_3D_image_slices}
\alias{plot_3D_image_slices}
\title{Plot 3D Images by Slices by Slices}
\usage{
plot_3D_image_slices(
  images,
  grids,
  Z_range,
  Z_slices = NULL,
  gridnames = c("X", "Y", "Z"),
  image_names = NULL
)
}
\arguments{
\item{images}{either an array for one image or a list of arrays for multiple images}

\item{grids}{a list of grids for each dimenion of the image.}

\item{Z_range}{lower and upper bound of Z slices}

\item{Z_slices}{a vector of Z slices}

\item{gridnames}{X, Y, Z axis names}

\item{image_names}{the name label for the images}
}
\description{
Plot 3D Images by Slices by Slices
}
\examples{
grids = lapply(1:3,function(i) seq(-29,29,by=1))
dim_img = sapply(1:3,function(i) length(grids[[i]]))
neighbors = find_image_neighbors(grids,radius=1)
sphere_mask = create_sphere_mask(grids,radius=20)
raw_img = ifelse(sphere_mask,rnorm(sum(sphere_mask)),NA)
smooth_img = spatial_kernel_smooth(raw_img,grids,sphere_mask,neighbors=neighbors,n_cor=0.5)
pg <- plot_3D_image_slices(images=list(raw=raw_img,smooth=smooth_img),
grids,Z_slices=seq(-18,18,by=9))
plot(pg)
}
\author{
Jian Kang <jiankang@umich.edu>
}
