% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCorrImage.R
\name{Spat_Corr_3D_images}
\alias{Spat_Corr_3D_images}
\title{Spatially varying correlation analysis between two groups of 3D images}
\usage{
Spat_Corr_3D_images(
  img_1,
  img_2,
  grids,
  mask = NULL,
  voxel_neighbors = NULL,
  rho_list = sort(c(0, seq(-0.95, 0.95, length = 100))),
  adj_dist = 1,
  size = 5,
  n_cor = 0.9,
  pos_prob = 0.8,
  neg_prob = 0.8,
  est_cormat = TRUE
)
}
\arguments{
\item{img_1}{a 4D array of multiple 3D images}

\item{img_2}{a 4D array of multiple 3D images}

\item{grids}{a list of grids for each dimenion of the image.}

\item{mask}{a 3D array of maske taking logicdal values. The default is NULL.}

\item{voxel_neighbors}{a list object contains two matrices: \code{idx} and \code{dist}.
\code{idx} contains the neighbor indices of each voxel in each row.
\code{dist} contains the distances between neightbors and the voxel in each row.}

\item{rho_list}{a vector of correlation values to be tested.}

\item{adj_dist}{the distance of adjacent nodes.}

\item{size}{the minimum clutersize.}

\item{n_cor}{the correlation between neighboring voxels for smoothing.}

\item{pos_prob}{positive correlation cluster threshold probablity.}

\item{neg_prob}{negative correlation cluster threshold probablity.}

\item{est_cormat}{TRUE or FALSE indicating that whether a minimum contrast method is used to estimate
the spatial correlations within each imaging modality. Default is TRUE}
}
\value{
a list of objects containing all the results
\describe{
\item{cor_img}{A correlation map}
\item{smooth_cor_est}{A smoothed correlation map}
\item{pos_cluster}{A logical image for all clusters with positive correlation}
\item{neg_cluster}{A logical image for all clusters with negative correlation}
\item{neg_group}{A list of voxel indices for all clusters with negative correlation}
\item{pos_group}{A list of voxel indices for all clusters with positive correlation}
\item{all_group}{A list of voxel indices for all clusters}
\item{simple_est_all}{A vector of mean correlation estimates within clusters}
\item{mle_all}{A list objects of the MLEs and likelihood ratio tests results}
\item{tab}{A summary of results}
\item{dat_1}{preprocessed images for \code{img_1} with each raw representing one image in vector format}
\item{dat_2}{preprocessed images for \code{img_2} with each raw representing one image in vector format}
\item{coords}{locations for all voxels}
}
}
\description{
This function is to estimate and test voxel-wise spatially varying correlation between two groups of 3D images
}
\examples{
set.seed(1000)
dim_img = c(10,10,10)
n = 50
grids <- lapply(1:3,function(i) seq(-round(dim_img[i]/2)+1,round(dim_img[i]/2),length=dim_img[i]))
img_1 <- array(rnorm(prod(dim_img)*n),dim=c(dim_img,n))
img_2 <- array(rnorm(prod(dim_img)*n),dim=c(dim_img,n))
cor_region <- create_sphere_mask(grids,radius=3)
cor_region_list <- array(cor_region,dim=dim(img_2))
img_2 <- ifelse(cor_region_list,img_1+rnorm(prod(dim_img)*n,sd=0.5),img_2)
mask <- create_sphere_mask(grids,radius=5)
cor_img <- comp_3D_images_corr(img_1,img_2,mask)
plot_3D_image_slices(cor_img,grids,c(-2,2))
res <- Spat_Corr_3D_images(img_1,img_2,grids,mask,pos_prob=0.90,neg_prob=0.90,n_cor=0.5,size=10)
plot_3D_image_slices(list(cor=res$cor_img,
smooth_cor=res$smooth_cor_est,
pos_cluster=res$pos_cluster),grids,c(-3,3))

}
\references{
Li L, Kang J, Lockhart SN, Adams J, Jagust WJ. (2019) Spatially adaptive varying correlation analysis for multimodal neuroimaging data. IEEE Trans Med Imaging. 38(1):113-123. PMCID: PMC6324929
}
\author{
Jian Kang <jiankang@umich.edu>
}
