% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCorrImage.R
\name{spatial_kernel_smooth}
\alias{spatial_kernel_smooth}
\title{Spatial adaptive kernel smoothing}
\usage{
spatial_kernel_smooth(val, grids, neighbors = NULL, radius = 3, n_cor = 0.9)
}
\arguments{
\item{\code{val}}{a n x 1 vector of imaging data.}

\item{\code{grids}}{a matrix of x, y, z coordinates.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Spatial adaptive kernel smoothing
}
\examples{
n = 1000
val = rnorm(n)
}
